+++ src/app/components/ImageUploaderExpanded/hooks/types.ts
@@ -1,14 +1,17 @@
 import { RefObject } from 'react';
 
 import { InitialFile } from '@/app/components/FileUploader/types';
+import { SelectedImage } from '../types';
 import { UseFileUploaderExpandedReturn } from '@/app/components/FileUploaderexpanded/hooks/types';
 
 export interface UseImageUploaderExpandedParams {
-  onImage: (file: File | null) => void;
+  onImage: (file: File | SelectedImage[] | null) => void;
   accept?: string;
   disabled?: boolean;
   placeholder?: string;
   initialFile?: InitialFile;
+  initialFiles?: SelectedImage[];
+  multiple?: boolean;
 }
 
 export interface UseImageUploaderExpandedReturn
@@ -19,4 +22,7 @@ export interface UseImageUploaderExpandedReturn
   openCamera: () => void;
   closeCamera: () => void;
   handleCaptureFromCamera: (file: File) => void;
+  images?: SelectedImage[];
+  toggleImage?: (id: string) => void;
+  clearImages?: () => void;
 }